+++ app/print_json/print_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func TestPrint(t *testing.T) {
-	probs, _ := probabilities.NewProbabilities(2, 5, 7, []int{1, 2, 3, 1})
+	probs, _ := probabilities.NewProbabilities(2, 5, 7, []uint64{1, 2, 3, 1})
 	ret := print_json.Print(probs, "Hello World")
 	var obj map[string]any
 	assert.NoError(t, json.Unmarshal([]byte(ret), &obj))
