+++ app/print_svg/print_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func TestPrint(t *testing.T) {
-	probs, _ := probabilities.NewProbabilities(2, 5, 7, []int{1, 2, 3, 1})
+	probs, _ := probabilities.NewProbabilities(2, 5, 7, []uint64{1, 2, 3, 1})
 	ret := print_svg.Print(probs, "Hello World")
 	assert.Contains(t, ret, ">Hello World</text>")
 	assert.Equal(t, strings.Count(ret, "<circle"), 4)
